+++ freeway-http/src/test/java/com/jujin/freeway/http/engine/FreewayHttpEngineTest.java
@@ -74,8 +74,8 @@ class FreewayHttpEngineTest {
         System.clearProperty(HttpConfigKeys.SSL_KEY_STORE_PASSWORD);
         System.clearProperty(HttpConfigKeys.SSL_KEY_STORE_TYPE);
         System.clearProperty(HttpConfigKeys.SSL_HTTP2);
-        System.clearProperty("freeway.web.server.port");
-        System.clearProperty("freeway.web.server.host");
+        System.clearProperty("freeway.http.server.port");
+        System.clearProperty("freeway.http.server.host");
     }
 
     @Test
@@ -310,8 +310,8 @@ class FreewayHttpEngineTest {
     @Test
     void servesRoutesWithLegacyWebKeys() throws Exception {
         int port = freePort();
-        System.setProperty("freeway.web.server.host", "127.0.0.1");
-        System.setProperty("freeway.web.server.port", String.valueOf(port));
+        System.setProperty("freeway.http.server.host", "127.0.0.1");
+        System.setProperty("freeway.http.server.port", String.valueOf(port));
 
         app = FreewayApp.run(new String[0], new PingModule());
         assertTrue(app.get(WebServer.class).isRunning());